Audio stuff moved around, removed MusicManager dependencies, eMusicMode used for ChangeMusicMode
