+++ src/peds/Ped.cpp
@@ -45,7 +45,6 @@
 #include "CarAI.h"
 #include "Zones.h"
 #include "Cranes.h"
-#include "MusicManager.h"
 #include "Timecycle.h"
 #include "ParticleObject.h"
 #include "Floater.h"
