+++ src/skel/win/win.cpp
@@ -2162,7 +2162,7 @@ _WinMain(HINSTANCE instance,
 		CPad::ResetCheats();
 		CPad::StopPadsShaking();
 		
-		DMAudio.ChangeMusicMode(_TODOCONST(3));
+		DMAudio.ChangeMusicMode(MUSICMODE_OFF);
 		
 		CTimer::Stop();
 		
@@ -2170,7 +2170,7 @@ _WinMain(HINSTANCE instance,
 		{
 			CGame::ShutDownForRestart();
 			CGame::InitialiseWhenRestarting();
-			DMAudio.ChangeMusicMode(_TODOCONST(1));
+			DMAudio.ChangeMusicMode(MUSICMODE_GAME);
 			LoadSplash(GetLevelSplashScreen(CGame::currLevel));
 			FrontEndMenuManager.m_bLoadingSavedGame = false;
 		}
